requirements.txt
fastapi
uvicorn[standard]
requests
pandas
sqlalchemy
pydantic
python-dotenv
apscheduler
numpy
fredapi